+++ src/arkreactor/Compiler/Macros/Executors/List.cpp
@@ -1,4 +1,5 @@
 #include <Ark/Compiler/Macros/Executors/List.hpp>
+#include <iostream>
 
 namespace Ark::internal
 {
@@ -21,12 +22,20 @@ namespace Ark::internal
             {
                 Node temp_body = macro->constList()[2];
                 Node args = macro->constList()[1];
+                std::size_t args_needed = args.list().size();
+                std::size_t args_given = node.constList().size() - 1;  // remove the first (the name of the macro)
+                std::string macro_name = macro->constList()[0].string();
+                bool has_spread = args.list().back().nodeType() == NodeType::Spread;
 
                 // bind node->list() to temp_body using macro->constList()[1]
                 std::unordered_map<std::string, Node> args_applied;
                 std::size_t j = 0;
                 for (std::size_t i = 1, end = node.constList().size(); i < end; ++i)
                 {
+                    // by breaking early if we have too many arguments, the args_applied/args_needed check will fail
+                    if (j >= args_needed)
+                        break;
+
                     const std::string& arg_name = args.list()[j].string();
                     if (args.list()[j].nodeType() == NodeType::Symbol)
                     {
@@ -46,22 +55,18 @@ namespace Ark::internal
                 }
 
                 // check argument count
-                if (args_applied.size() + 1 == args.list().size() && args.list().back().nodeType() == NodeType::Spread)
+                if (args_applied.size() + 1 == args_needed && has_spread)
                 {
                     // just a spread we didn't assign
                     args_applied[args.list().back().string()] = Node(NodeType::List);
                     args_applied[args.list().back().string()].push_back(Node::getListNode());
                 }
-                else if (args_applied.size() != args.list().size())
-                {
-                    std::size_t args_needed = args.list().size();
-                    std::string macro_name = macro->constList()[0].string();
 
-                    if (args.list().back().nodeType() != NodeType::Spread)
-                        throwMacroProcessingError("Macro `" + macro_name + "' got " + std::to_string(args_applied.size()) + " argument(s) but needed " + std::to_string(args_needed), *macro);
-                    else
-                        throwMacroProcessingError("Macro `" + macro_name + "' got " + std::to_string(args_applied.size()) + " argument(s) but needed at least " + std::to_string(args_needed - 1), *macro);
-                }
+                if (args_given != args_needed && !has_spread)
+                    throwMacroProcessingError("Macro `" + macro_name + "' got " + std::to_string(args_given) + " argument(s) but needed " + std::to_string(args_needed), node);
+                else if (args_applied.size() != args_needed && has_spread)
+                    // args_needed - 1 because we do not count the spread as a required argument
+                    throwMacroProcessingError("Macro `" + macro_name + "' got " + std::to_string(args_applied.size()) + " argument(s) but needed at least " + std::to_string(args_needed - 1), node);
 
                 if (!args_applied.empty())
                     unify(args_applied, temp_body, nullptr);